Finished Ability To Compile Circuits
You can now create new circuits by giving it a name and pressing the create button. This will save a snapshot of the name and the compiled logic behind the circuit. Then a button will be added to the UI that will create the new Circuit with its newly compiled logic when it is clicked.